make scripts/archive_dbs.sh respect lock
and skip archiving when a corresponding .lock file is present


Former-commit-id: f1e4df0d89d8fd8c3937a6612882adfd580e9cbd

+++ scripts/archive_dbs.sh
@@ -25,6 +25,12 @@ do
     lock="${line}.lock"
     gz="${line}.gz"
 
+    if [ -f ${lock} ]
+    then
+        echo >&2 "skipping due to ${lock}"
+        continue
+    fi
+
     touch ${lock}
     gzip ${db}
     rm ${lock}
